+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Improve MapperUtils.convertValue 
+- Add MapperUtils.readValue for java class 
+
 ## [3.0.0] - 2023-03-24
 
 ### Added